@@ -28,93 +31,97 @@
         "test:e2e": "NODE_ENV=test jest --config ./test/jest-e2e.json",
         "docs:dev": "vuepress dev -p 7070",
         "docs:build": "DEPLOY_ENV=gh-pages vuepress build",
-        "docs:deploy": "yarn docs:build && gh-pages -d .vuepress/dist"
+        "docs:deploy": "yarn docs:build && gh-pages -d .vuepress/dist",
+        "prepare": "husky install"
     },
     "dependencies": {
         "@godaddy/terminus": "^4.9.0",
-        "@nestjs/common": "^7.6.17",
-        "@nestjs/config": "^0.6.3",
-        "@nestjs/core": "^7.6.17",
-        "@nestjs/jwt": "^7.2.0",
-        "@nestjs/microservices": "^7.6.17",
-        "@nestjs/passport": "^7.1.5",
-        "@nestjs/platform-express": "^7.6.17",
-        "@nestjs/swagger": "^4.8.0",
-        "@nestjs/terminus": "^7.2.0",
-        "@nestjs/typeorm": "^7.1.5",
-        "@types/lodash": "^4.14.170",
-        "aws-sdk": "^2.922.0",
+        "@nestjs/axios": "^0.0.1",
+        "@nestjs/common": "^8.0.6",
+        "@nestjs/config": "^1.0.1",
+        "@nestjs/core": "^8.0.6",
+        "@nestjs/jwt": "^8.0.0",
+        "@nestjs/microservices": "^8.0.6",
+        "@nestjs/passport": "^8.0.1",
+        "@nestjs/platform-express": "^8.0.6",
+        "@nestjs/swagger": "^5.0.9",
+        "@nestjs/terminus": "^8.0.0",
+        "@nestjs/typeorm": "^8.0.2",
+        "aws-sdk": "^2.989.0",
         "bcrypt": "^5.0.1",
         "class-transformer": "^0.4.0",
         "class-validator": "^0.13.1",
         "compression": "^1.7.4",
         "express": "^4.17.1",
-        "express-rate-limit": "^5.2.6",
-        "file-type": "^16.5.0",
+        "express-rate-limit": "^5.3.0",
+        "file-type": "^16.5.3",
         "helmet": "^4.6.0",
         "jsonwebtoken": "^8.5.1",
         "lodash": "^4.17.21",
-        "mime-types": "^2.1.31",
+        "mime-types": "^2.1.32",
         "morgan": "^1.10.0",
-        "nats": "^1.4.12",
-        "nestjs-i18n": "^8.1.1",
+        "nats": "^2.2.0",
+        "nestjs-i18n": "^8.1.6",
         "passport": "^0.4.1",
         "passport-jwt": "^4.0.0",
-        "pg": "^8.6.0",
+        "pg": "^8.7.1",
         "reflect-metadata": "^0.1.13",
         "request-context": "^2.0.0",
-        "rxjs": "^6.0.0",
-        "source-map-support": "^0.5.19",
+        "rxjs": "^7.3.0",
+        "source-map-support": "^0.5.20",
         "swagger-ui-express": "^4.1.6",
-        "typeorm": "^0.2.34",
+        "typeorm": "^0.2.37",
         "typeorm-transactional-cls-hooked": "^0.1.21",
         "uuid": "^8.3.2"
     },
     "devDependencies": {
-        "@nestjs/testing": "^7.6.17",
+        "@nestjs/cli": "^8.1.1",
+        "@nestjs/testing": "^8.0.6",
         "@types/bcrypt": "^5.0.0",
-        "@types/compression": "^1.7.0",
-        "@types/express": "^4.17.12",
-        "@types/express-rate-limit": "^5.1.1",
-        "@types/file-type": "^10.9.1",
-        "@types/helmet": "^4.0.0",
-        "@types/jest": "^26.0.23",
-        "@types/jsonwebtoken": "^8.5.1",
-        "@types/mime-types": "^2.1.0",
-        "@types/morgan": "^1.9.2",
-        "@types/node": "^15.12.1",
-        "@types/passport-jwt": "^3.0.5",
+        "@types/compression": "^1.7.2",
+        "@types/express": "^4.17.13",
+        "@types/express-rate-limit": "^5.1.3",
+        "@types/faker": "^5.5.8",
+        "@types/jest": "^27.0.1",
+        "@types/jsonwebtoken": "^8.5.5",
+        "@types/lodash": "^4.14.172",
+        "@types/mime-types": "^2.1.1",
+        "@types/morgan": "^1.9.3",
+        "@types/node": "^16.9.1",
+        "@types/passport-jwt": "^3.0.6",
         "@types/supertest": "^2.0.11",
-        "@types/uuid": "^8.3.0",
-        "@types/webpack-env": "^1.16.0",
-        "@typescript-eslint/eslint-plugin": "^4.26.0",
-        "@typescript-eslint/eslint-plugin-tslint": "^4.26.0",
-        "@typescript-eslint/parser": "^4.26.0",
-        "clean-webpack-plugin": "^4.0.0-alpha.0",
+        "@types/uuid": "^8.3.1",
+        "@types/webpack-env": "^1.16.2",
+        "@typescript-eslint/eslint-plugin": "^4.31.1",
+        "@typescript-eslint/eslint-plugin-tslint": "^4.31.1",
+        "@typescript-eslint/parser": "^4.31.1",
+        "clean-webpack-plugin": "^4.0.0",
         "copyfiles": "^2.4.1",
         "cross-env": "^7.0.3",
-        "eslint": "^7.28.0",
+        "eslint": "^7.32.0",
         "eslint-config-prettier": "^8.3.0",
-        "eslint-plugin-import": "^2.23.4",
+        "eslint-plugin-import": "^2.24.2",
         "eslint-plugin-import-helpers": "^1.1.0",
-        "eslint-plugin-prettier": "^3.4.0",
+        "eslint-plugin-prettier": "^4.0.0",
         "eslint-plugin-simple-import-sort": "^7.0.0",
-        "eslint-plugin-unicorn": "^33.0.1",
-        "gh-pages": "^3.2.0",
-        "husky": "^6.0.0",
-        "jest": "^27.0.4",
-        "lint-staged": "~11.0.0",
-        "nodemon": "^2.0.7",
-        "prettier": "^2.3.1",
-        "supertest": "^6.1.3",
-        "ts-jest": "^27.0.3",
-        "ts-loader": "^9.2.3",
-        "ts-node": "^10.0.0",
+        "eslint-plugin-sonarjs": "^0.10.0",
+        "eslint-plugin-unicorn": "^36.0.0",
+        "gh-pages": "^3.2.3",
+        "husky": "^7.0.0",
+        "jest": "^27.2.0",
+        "lint-staged": "~11.1.2",
+        "nodemon": "^2.0.12",
+        "prettier": "^2.4.0",
+        "supertest": "^6.1.6",
+        "ts-jest": "^27.0.5",
+        "ts-loader": "^9.2.5",
+        "ts-node": "^10.2.1",
         "tslint": "^6.1.3",
-        "typescript": "^4.3.2",
+        "typeorm-seeding": "^1.6.1",
+        "typescript": "^4.4.3",
         "vuepress": "^1.8.2",
-        "webpack": "^5.38.1",
-        "webpack-cli": "^4.7.0",
+        "webpack": "^5.52.1",
+        "webpack-cli": "^4.8.0",
         "webpack-node-externals": "^3.0.0"
     },
     "jest": {
