@@ -1,12 +1,14 @@
-import { Inject, Injectable } from '@nestjs/common';
+import { Inject, Injectable, Optional } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
 import type { HealthIndicatorResult } from '@nestjs/terminus';
 import { HealthCheckError, HealthIndicator } from '@nestjs/terminus';
+import { firstValueFrom } from 'rxjs';
 import { timeout } from 'rxjs/operators';
 
 @Injectable()
 export class ServiceHealthIndicator extends HealthIndicator {
   constructor(
+    @Optional()
     @Inject('NATS_SERVICE')
     private readonly clientProxy: ClientProxy,
   ) {
@@ -15,10 +17,20 @@ export class ServiceHealthIndicator extends HealthIndicator {
 
   async isHealthy(eventName: string): Promise<HealthIndicatorResult> {
     try {
-      const result = await this.clientProxy
-        .send(eventName, { check: true })
-        .pipe(timeout(10_000))
-        .toPromise();
+      if (!this.clientProxy) {
+        return {
+          [eventName]: {
+            status: 'down',
+          },
+        };
+      }
+
+      const result = await firstValueFrom(
+        this.clientProxy.send(eventName, { check: true }).pipe(timeout(10_000)),
+        {
+          defaultValue: undefined,
+        },
+      );
 
       return {
         [eventName]: result,
