@@ -1,356 +1,403 @@
 module.exports = {
-    parser: '@typescript-eslint/parser',
-    env: {
-        jest: true,
-        es6: true,
-        node: true,
-    },
-    ignorePatterns: ['.eslintrc.js'],
-    parserOptions: {
-        project: './tsconfig.json',
-        sourceType: 'module',
-    },
-    extends: [
-        'plugin:import/errors',
-        'plugin:import/warnings',
-        'plugin:import/typescript',
-        'plugin:prettier/recommended',
-        'eslint:recommended',
-        'plugin:@typescript-eslint/recommended',
-        'plugin:@typescript-eslint/recommended-requiring-type-checking',
-        'plugin:@typescript-eslint/eslint-recommended',
+  parser: '@typescript-eslint/parser',
+  env: {
+    jest: true,
+    es6: true,
+    node: true,
+  },
+  ignorePatterns: ['.eslintrc.js'],
+  parserOptions: {
+    ecmaVersion: 2018,
+    project: './tsconfig.eslint.json',
+    sourceType: 'module',
+  },
+  extends: [
+    'plugin:import/errors',
+    'plugin:import/warnings',
+    'plugin:import/typescript',
+    'plugin:prettier/recommended',
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended',
+    'plugin:@typescript-eslint/recommended-requiring-type-checking',
+    'plugin:@typescript-eslint/eslint-recommended',
+    'plugin:sonarjs/recommended',
+  ],
+  plugins: [
+    '@typescript-eslint',
+    '@typescript-eslint/tslint',
+    'prettier',
+    'simple-import-sort',
+    'import',
+    'unicorn',
+    'sonarjs',
+  ],
+  rules: {
+    'sonarjs/cognitive-complexity': 'error',
+    'sonarjs/no-identical-expressions': 'error',
+    'sonarjs/no-duplicate-string': 'off',
+    'simple-import-sort/imports': 'error',
+    'import/newline-after-import': 'error',
+    'import/no-absolute-path': 'error',
+    'import/no-duplicates': 'error',
+    'prettier/prettier': ['error', { singleQuote: true, trailingComma: 'all', tabWidth: 2 }],
+    '@typescript-eslint/camelcase': 'off',
+    '@typescript-eslint/no-unused-vars': 'off',
+    '@typescript-eslint/no-unused-vars-experimental': [
+      'error',
+      { ignoredNamesRegex: '^_' },
     ],
-    plugins: [
-        '@typescript-eslint',
-        '@typescript-eslint/tslint',
-        'prettier',
-        'simple-import-sort',
-        'import',
-        'unicorn',
+    '@typescript-eslint/adjacent-overload-signatures': 'error',
+    'no-await-in-loop': 'error',
+    'padding-line-between-statements': [
+      'error',
+      { blankLine: 'always', prev: '*', next: 'return' },
+      { blankLine: 'always', prev: '*', next: 'try' },
+      { blankLine: 'always', prev: 'try', next: '*' },
+      { blankLine: 'always', prev: '*', next: 'block-like' },
+      { blankLine: 'always', prev: 'block-like', next: '*' },
+      { blankLine: 'always', prev: '*', next: 'throw' },
+      { blankLine: 'always', prev: 'var', next: '*' },
     ],
-    rules: {
-        'simple-import-sort/imports': 'error',
-        'prettier/prettier': ['error', { singleQuote: true, trailingComma: 'all', tabWidth: 2 }],
-        '@typescript-eslint/camelcase': 'off',
-        '@typescript-eslint/no-unused-vars-experimental': [
-            'error',
-            { ignoredNamesRegex: '^_' },
+    '@typescript-eslint/array-type': ['error', { default: 'array-simple' }],
+    '@typescript-eslint/ban-types': [
+      'error',
+      {
+        types: {
+          'Object': {
+            message: 'Avoid using the `Object` type. Did you mean `object`?',
+          },
+          'Function': {
+            message: 'Avoid using the `Function` type. Prefer a specific function type, like `() => void`.',
+          },
+          'Boolean': {
+            message: 'Avoid using the `Boolean` type. Did you mean `boolean`?',
+            fixWith: 'boolean',
+          },
+          'Number': {
+            message: 'Avoid using the `Number` type. Did you mean `number`?',
+            fixWith: 'number',
+          },
+          'Symbol': {
+            message: 'Avoid using the `Symbol` type. Did you mean `symbol`?',
+            fixWith: 'symbol',
+          },
+          'String': {
+            message: 'Avoid using the `String` type. Did you mean `string`?',
+            fixWith: 'string',
+          },
+          '{}': {
+            message: 'Use Record<K, V> instead',
+            fixWith: 'Record<K, V>',
+          },
+          'object': {
+            message: 'Use Record<K, V> instead',
+            fixWith: 'Record<K, V>',
+          },
+        },
+      },
+    ],
+    '@typescript-eslint/explicit-member-accessibility': [
+      'off',
+      {
+        overrides: {
+          constructors: 'off',
+        },
+      },
+    ],
+    // '@typescript-eslint/indent': 'off',
+    '@typescript-eslint/member-delimiter-style': [
+      'error',
+      {
+        multiline: {
+          delimiter: 'semi',
+          requireLast: true,
+        },
+        singleline: {
+          delimiter: 'semi',
+          requireLast: false,
+        },
+      },
+    ],
+    '@typescript-eslint/member-ordering': 'off',
+    '@typescript-eslint/no-angle-bracket-type-assertion': 'off',
+    '@typescript-eslint/no-empty-function': 'error',
+    '@typescript-eslint/no-empty-interface': 'error',
+    '@typescript-eslint/no-explicit-any': 'error',
+    '@typescript-eslint/no-inferrable-types': 'error',
+    '@typescript-eslint/await-thenable': 'error',
+    '@typescript-eslint/ban-ts-comment': 'error',
+    '@typescript-eslint/ban-tslint-comment': 'error',
+    '@typescript-eslint/consistent-indexed-object-style': 'error',
+    '@typescript-eslint/consistent-type-definitions': ['error', 'interface'],
+    '@typescript-eslint/consistent-type-imports': [
+      'error',
+      { prefer: 'type-imports' },
+    ],
+    '@typescript-eslint/no-misused-new': 'error',
+    '@typescript-eslint/no-unsafe-assignment': 'off',
+    '@typescript-eslint/no-unsafe-call': 'off',
+    '@typescript-eslint/no-unsafe-member-access': 'off',
+    '@typescript-eslint/no-unsafe-return': 'off',
+    '@typescript-eslint/restrict-template-expressions': 'off',
+    '@typescript-eslint/no-require-imports': 'error',
+    'keyword-spacing': 'off',
+    '@typescript-eslint/keyword-spacing': 'error',
+    '@typescript-eslint/no-namespace': 'error',
+    '@typescript-eslint/no-this-alias': 'error',
+    '@typescript-eslint/no-use-before-define': 'error',
+    '@typescript-eslint/no-use-before-declare': 'off',
+    '@typescript-eslint/no-var-requires': 'error',
+    '@typescript-eslint/prefer-for-of': 'error',
+    '@typescript-eslint/prefer-function-type': 'error',
+    '@typescript-eslint/prefer-namespace-keyword': 'error',
+    '@typescript-eslint/quotes': [
+      'error',
+      'single',
+      {
+        avoidEscape: true,
+      },
+    ],
+    '@typescript-eslint/semi': ['error', 'always'],
+    '@typescript-eslint/naming-convention': [
+      'error',
+      {
+        selector: 'default',
+        format: ['camelCase', 'PascalCase', 'snake_case', 'UPPER_CASE'],
+        filter: {
+          regex: '^_.*$',
+          match: false,
+        },
+      },
+      {
+        selector: 'variable',
+        format: ['camelCase', 'UPPER_CASE'],
+      },
+      {
+        selector: 'interface',
+        format: ['PascalCase'],
+        prefix: ['I'],
+      },
+      {
+        selector: 'typeLike',
+        format: ['PascalCase'],
+      },
+      {
+        selector: 'memberLike',
+        modifiers: ['private'],
+        format: ['camelCase'],
+        leadingUnderscore: 'forbid',
+      },
+      {
+        selector: 'variable',
+        types: ['boolean'],
+        format: ['PascalCase'],
+        prefix: ['is', 'should', 'has', 'can', 'did', 'will'],
+      },
+    ],
+    '@typescript-eslint/type-annotation-spacing': 'error',
+    '@typescript-eslint/unified-signatures': 'error',
+    'arrow-body-style': 'error',
+    'arrow-parens': ['error', 'always'],
+    'camelcase': 'off',
+    'complexity': 'off',
+    'no-restricted-imports': [
+      'error',
+      {
+        paths: [
+          {
+            name: 'rxjs/Rx',
+            message: 'Please import directly from \'rxjs\' instead',
+          },
         ],
-        '@typescript-eslint/adjacent-overload-signatures': 'error',
-        '@typescript-eslint/array-type': ['error', { default: 'array-simple' }],
-        '@typescript-eslint/ban-types': [
-            'error',
-            {
-                types: {
-                    'String': {
-                        message: 'Use string instead',
-                        fixWith: 'string',
-                    },
-                    '{}': {
-                        message: 'Use Record<K, V> instead',
-                        fixWith: 'Record<K, V>',
-                    },
-                    'object': {
-                        message: 'Use Record<K, V> instead',
-                        fixWith: 'Record<K, V>',
-                    },
-                },
-            },
+      },
+    ],
+    '@typescript-eslint/interface-name-prefix': 'off',
+    '@typescript-eslint/no-non-null-assertion': 'error',
+    'object-curly-spacing': ['error', 'always'],
+    'no-multi-spaces': 'error',
+    'no-useless-return': 'error',
+    'no-else-return': 'error',
+    'no-implicit-coercion': 'error',
+    'constructor-super': 'error',
+    'yoda': 'error',
+    'strict': ['error', 'never'],
+    'curly': 'error',
+    'dot-notation': 'error',
+    'eol-last': 'error',
+    'eqeqeq': ['error', 'smart'],
+    'guard-for-in': 'error',
+    'id-match': 'error',
+    'import/no-default-export': 'error',
+    'import/no-deprecated': 'warn',
+    'import/no-internal-modules': 'off',
+    'import/order': 'off',
+    'max-classes-per-file': ['error', 1],
+    'max-len': [
+      'error',
+      {
+        code: 150,
+      },
+    ],
+    'new-parens': 'error',
+    'no-bitwise': 'error',
+    'no-caller': 'error',
+    'no-cond-assign': 'error',
+    'no-constant-condition': 'error',
+    'no-dupe-else-if': 'error',
+    'lines-between-class-members': ['error', 'always'],
+    'no-console': [
+      'error',
+      {
+        allow: [
+          'info',
+          'dirxml',
+          'warn',
+          'error',
+          'dir',
+          'timeLog',
+          'assert',
+          'clear',
+          'count',
+          'countReset',
+          'group',
+          'groupCollapsed',
+          'groupEnd',
+          'table',
+          'Console',
+          'markTimeline',
+          'profile',
+          'profileEnd',
+          'timeline',
+          'timelineEnd',
+          'timeStamp',
+          'context',
         ],
-        '@typescript-eslint/explicit-member-accessibility': [
-            'off',
-            {
-                overrides: {
-                    constructors: 'off',
-                },
-            },
-        ],
-        // '@typescript-eslint/indent': 'off',
-        '@typescript-eslint/member-delimiter-style': [
-            'error',
-            {
-                multiline: {
-                    delimiter: 'semi',
-                    requireLast: true,
-                },
-                singleline: {
-                    delimiter: 'semi',
-                    requireLast: false,
-                },
-            },
-        ],
-        '@typescript-eslint/member-ordering': 'off',
-        '@typescript-eslint/no-angle-bracket-type-assertion': 'off',
-        '@typescript-eslint/no-empty-function': 'error',
-        '@typescript-eslint/no-empty-interface': 'error',
-        '@typescript-eslint/no-explicit-any': 'error',
-        '@typescript-eslint/no-inferrable-types': 'error',
-        '@typescript-eslint/await-thenable': 'error',
-        '@typescript-eslint/ban-ts-comment': 'error',
-        '@typescript-eslint/ban-tslint-comment': 'error',
-        '@typescript-eslint/consistent-indexed-object-style': 'error',
-        '@typescript-eslint/consistent-type-definitions': ['error', 'interface'],
-        '@typescript-eslint/consistent-type-imports': [
-            'error',
-            { prefer: 'type-imports' },
-        ],
-        '@typescript-eslint/no-misused-new': 'error',
-        '@typescript-eslint/no-unsafe-assignment': 'off',
-        '@typescript-eslint/no-unsafe-call': 'off',
-        '@typescript-eslint/no-unsafe-member-access': 'off',
-        '@typescript-eslint/no-unsafe-return': 'off',
-        '@typescript-eslint/restrict-template-expressions': 'off',
-        '@typescript-eslint/no-require-imports': 'error',
-        'keyword-spacing': 'off',
-        '@typescript-eslint/keyword-spacing': 'error',
-        '@typescript-eslint/no-namespace': 'error',
-        '@typescript-eslint/no-this-alias': 'error',
-        '@typescript-eslint/no-use-before-define': 'error',
-        '@typescript-eslint/no-use-before-declare': 'off',
-        '@typescript-eslint/no-var-requires': 'error',
-        '@typescript-eslint/prefer-for-of': 'error',
-        '@typescript-eslint/prefer-function-type': 'error',
-        '@typescript-eslint/prefer-namespace-keyword': 'error',
-        '@typescript-eslint/quotes': [
-            'error',
-            'single',
-            {
-                avoidEscape: true,
-            },
-        ],
-        '@typescript-eslint/semi': ['error', 'always'],
-        '@typescript-eslint/naming-convention': [
-            'error',
-            {
-                selector: 'default',
-                format: ['camelCase', 'PascalCase', 'snake_case', 'UPPER_CASE'],
-                filter: {
-                    regex: '^_.*$',
-                    match: false,
-                },
-            },
-            {
-                selector: 'variable',
-                format: ['camelCase', 'UPPER_CASE'],
-            },
-            {
-                selector: 'interface',
-                format: ['PascalCase'],
-                prefix: ['I'],
-            },
-            {
-                selector: 'typeLike',
-                format: ['PascalCase'],
-            },
-            {
-                selector: 'memberLike',
-                modifiers: ['private'],
-                format: ['camelCase'],
-                leadingUnderscore: 'forbid',
-            },
-            {
-                selector: 'variable',
-                types: ['boolean'],
-                format: ['PascalCase'],
-                prefix: ['is', 'should', 'has', 'can', 'did', 'will'],
-            },
-        ],
-        '@typescript-eslint/type-annotation-spacing': 'error',
-        '@typescript-eslint/unified-signatures': 'error',
-        'arrow-body-style': 'error',
-        'arrow-parens': ['error', 'always'],
-        'camelcase': 'off',
-        'complexity': 'off',
-        'no-restricted-imports': [
-            'error',
-            {
-                paths: [
-                    {
-                        name: 'rxjs/Rx',
-                        message: "Please import directly from 'rxjs' instead",
-                    },
-                ],
-            },
-        ],
-        '@typescript-eslint/interface-name-prefix': 'off',
-        '@typescript-eslint/no-non-null-assertion': 'error',
-        'object-curly-spacing': ['error', 'always'],
-        'no-multi-spaces': 'error',
-        'constructor-super': 'error',
-        'curly': 'error',
-        'dot-notation': 'error',
-        'eol-last': 'error',
-        'eqeqeq': ['error', 'smart'],
-        'guard-for-in': 'error',
-        'id-match': 'error',
-        'import/no-default-export': 'error',
-        'import/no-deprecated': 'warn',
-        'import/no-internal-modules': 'off',
-        'import/order': 'off',
-        'max-classes-per-file': ['error', 1],
-        'max-len': [
-            'error',
-            {
-                code: 150,
-            },
-        ],
-        'new-parens': 'error',
-        'no-bitwise': 'error',
-        'no-caller': 'error',
-        'no-cond-assign': 'error',
-        'no-console': [
-            'error',
-            {
-                allow: [
-                    'info',
-                    'dirxml',
-                    'warn',
-                    'error',
-                    'dir',
-                    'timeLog',
-                    'assert',
-                    'clear',
-                    'count',
-                    'countReset',
-                    'group',
-                    'groupCollapsed',
-                    'groupEnd',
-                    'table',
-                    'Console',
-                    'markTimeline',
-                    'profile',
-                    'profileEnd',
-                    'timeline',
-                    'timelineEnd',
-                    'timeStamp',
-                    'context',
-                ],
-            },
-        ],
-        'no-debugger': 'error',
-        'no-duplicate-case': 'error',
-        'no-duplicate-imports': 'off',
-        '@typescript-eslint/no-duplicate-imports': ['error'],
-        'no-empty': 'error',
-        'no-eval': 'error',
-        'no-extra-bind': 'error',
-        'no-fallthrough': 'error',
-        'no-invalid-this': 'error',
-        'no-irregular-whitespace': 'error',
-        'no-multiple-empty-lines': [
-            'error',
-            {
-                max: 1,
-            },
-        ],
-        'no-new-func': 'error',
-        'no-new-wrappers': 'error',
-        'no-redeclare': 'off',
-        'no-return-await': 'error',
-        'no-sequences': 'error',
-        'no-sparse-arrays': 'error',
-        'no-template-curly-in-string': 'error',
-        'no-shadow': 'off',
-        '@typescript-eslint/no-shadow': 'error',
-        'no-throw-literal': 'error',
-        'no-trailing-spaces': 'error',
-        'no-undef-init': 'error',
-        'no-unsafe-finally': 'error',
-        'no-unused-expressions': 'error',
-        'no-unused-labels': 'error',
-        'no-var': 'error',
-        'object-shorthand': 'error',
-        'prefer-const': 'error',
-        'prefer-object-spread': 'error',
-        'quote-props': ['error', 'consistent-as-needed'],
-        'radix': 'error',
-        'use-isnan': 'error',
-        'valid-typeof': 'off',
-        'space-before-function-paren': 'off',
-        'unicorn/better-regex': 'error',
-        'unicorn/catch-error-name': 'error',
-        'unicorn/consistent-destructuring': 'error',
-        'unicorn/consistent-function-scoping': 'error',
-        'unicorn/custom-error-definition': 'off',
-        'unicorn/empty-brace-spaces': 'error',
-        'unicorn/error-message': 'error',
-        'unicorn/escape-case': 'error',
-        'unicorn/expiring-todo-comments': 'error',
-        'unicorn/explicit-length-check': 'error',
-        'unicorn/filename-case': [
-            'error',
-            {
-                cases: {
-                    kebabCase: true,
-                    pascalCase: true
-                }
-            }
-        ],
-        'unicorn/import-index': 'error',
-        'unicorn/import-style': 'error',
-        'unicorn/new-for-builtins': 'error',
-        'unicorn/no-abusive-eslint-disable': 'error',
-        'unicorn/no-array-callback-reference': 'error',
-        'unicorn/no-array-for-each': 'error',
-        'unicorn/no-array-push-push': 'error',
-        'unicorn/no-array-reduce': 'error',
-        'unicorn/no-console-spaces': 'error',
-        'unicorn/no-document-cookie': 'error',
-        'unicorn/no-for-loop': 'error',
-        'unicorn/no-hex-escape': 'error',
-        'unicorn/no-instanceof-array': 'error',
-        'unicorn/no-keyword-prefix': 'off',
-        'unicorn/no-lonely-if': 'error',
-        'no-nested-ternary': 'off',
-        'unicorn/no-nested-ternary': 'error',
-        'unicorn/no-new-array': 'error',
-        'unicorn/no-new-buffer': 'error',
-        'unicorn/no-null': 'error',
-        'unicorn/no-object-as-default-parameter': 'error',
-        'unicorn/no-process-exit': 'error',
-        'unicorn/no-static-only-class': 'off',
-        'unicorn/no-this-assignment': 'error',
-        'unicorn/no-unreadable-array-destructuring': 'error',
-        'unicorn/no-unsafe-regex': 'off',
-        'unicorn/no-unused-properties': 'off',
-        'unicorn/no-useless-undefined': 'error',
-        'unicorn/no-zero-fractions': 'error',
-        'unicorn/number-literal-case': 'error',
-        'unicorn/numeric-separators-style': 'error',
-        'unicorn/prefer-add-event-listener': 'error',
-        'unicorn/prefer-array-find': 'error',
-        'unicorn/prefer-array-flat': 'error',
-        'unicorn/prefer-array-flat-map': 'error',
-        'unicorn/prefer-array-index-of': 'error',
-        'unicorn/prefer-array-some': 'error',
-        'unicorn/prefer-date-now': 'error',
-        'unicorn/prefer-default-parameters': 'error',
-        'unicorn/prefer-dom-node-append': 'error',
-        'unicorn/prefer-dom-node-dataset': 'error',
-        'unicorn/prefer-dom-node-remove': 'error',
-        'unicorn/prefer-dom-node-text-content': 'error',
-        'unicorn/prefer-includes': 'error',
-        'unicorn/prefer-keyboard-event-key': 'error',
-        'unicorn/prefer-math-trunc': 'error',
-        'unicorn/prefer-modern-dom-apis': 'error',
-        'unicorn/prefer-module': 'off',
-        'unicorn/prefer-negative-index': 'error',
-        'unicorn/prefer-node-protocol': 'off',
-        'unicorn/prefer-number-properties': 'error',
-        'unicorn/prefer-optional-catch-binding': 'error',
-        'unicorn/prefer-reflect-apply': 'error',
-        'unicorn/prefer-regexp-test': 'error',
-        'unicorn/prefer-set-has': 'error',
-        'unicorn/prefer-spread': 'error',
-        'unicorn/prefer-string-replace-all': 'off',
-        'unicorn/prefer-string-slice': 'error',
-        'unicorn/prefer-string-starts-ends-with': 'error',
-        'unicorn/prefer-string-trim-start-end': 'error',
-        'unicorn/prefer-switch': 'error',
-        'unicorn/prefer-ternary': 'error',
-        'unicorn/prefer-type-error': 'error',
-        'unicorn/prevent-abbreviations': 'off',
-        'unicorn/string-content': 'off',
-        'unicorn/throw-new-error': 'error',
-    },
+      },
+    ],
+    'no-debugger': 'error',
+    'no-duplicate-case': 'error',
+    'no-duplicate-imports': 'off',
+    '@typescript-eslint/no-duplicate-imports': ['error'],
+    'no-empty': 'error',
+    'no-eval': 'error',
+    'no-extra-bind': 'error',
+    'no-fallthrough': 'error',
+    'no-invalid-this': 'error',
+    'no-irregular-whitespace': 'error',
+    'no-multiple-empty-lines': [
+      'error',
+      {
+        max: 1,
+      },
+    ],
+    'no-new-func': 'error',
+    'no-new-wrappers': 'error',
+    'no-redeclare': 'error',
+    'no-return-await': 'error',
+    'no-sequences': 'error',
+    'no-sparse-arrays': 'error',
+    'no-template-curly-in-string': 'error',
+    'no-shadow': 'off',
+    '@typescript-eslint/no-shadow': 'error',
+    'no-throw-literal': 'error',
+    'no-trailing-spaces': 'error',
+    'no-undef-init': 'error',
+    'no-unsafe-finally': 'error',
+    'no-unused-expressions': 'error',
+    'no-unused-labels': 'error',
+    'no-var': 'error',
+    'object-shorthand': 'error',
+    'prefer-const': 'error',
+    'prefer-object-spread': 'error',
+    'quote-props': ['error', 'consistent-as-needed'],
+    'radix': 'error',
+    'use-isnan': 'error',
+    'valid-typeof': 'off',
+    'space-before-function-paren': 'off',
+    'unicorn/better-regex': 'error',
+    'unicorn/catch-error-name': 'error',
+    'unicorn/consistent-destructuring': 'error',
+    'unicorn/consistent-function-scoping': 'error',
+    'unicorn/custom-error-definition': 'off',
+    'unicorn/empty-brace-spaces': 'error',
+    'unicorn/error-message': 'error',
+    'unicorn/escape-case': 'error',
+    'unicorn/expiring-todo-comments': 'error',
+    'unicorn/explicit-length-check': 'error',
+    'unicorn/filename-case': [
+      'error',
+      {
+        cases: {
+          kebabCase: true,
+          pascalCase: true,
+        },
+      },
+    ],
+    'unicorn/import-index': 'error',
+    'unicorn/import-style': 'error',
+    'unicorn/new-for-builtins': 'error',
+    'unicorn/no-abusive-eslint-disable': 'error',
+    'unicorn/no-array-callback-reference': 'error',
+    'unicorn/no-array-for-each': 'error',
+    'unicorn/no-array-push-push': 'error',
+    'unicorn/no-array-reduce': 'error',
+    'unicorn/no-console-spaces': 'error',
+    'unicorn/no-document-cookie': 'error',
+    'unicorn/no-for-loop': 'error',
+    'unicorn/no-hex-escape': 'error',
+    'unicorn/no-instanceof-array': 'error',
+    'unicorn/no-keyword-prefix': 'off',
+    'unicorn/no-lonely-if': 'error',
+    'no-nested-ternary': 'off',
+    'unicorn/no-nested-ternary': 'error',
+    'unicorn/no-new-array': 'error',
+    'unicorn/no-new-buffer': 'error',
+    'unicorn/no-null': 'error',
+    'unicorn/no-object-as-default-parameter': 'error',
+    'unicorn/no-process-exit': 'error',
+    'unicorn/no-static-only-class': 'off',
+    'unicorn/no-this-assignment': 'error',
+    'unicorn/no-unreadable-array-destructuring': 'error',
+    'unicorn/no-unsafe-regex': 'off',
+    'unicorn/no-unused-properties': 'off',
+    'unicorn/no-useless-undefined': 'error',
+    'unicorn/no-zero-fractions': 'error',
+    'unicorn/number-literal-case': 'error',
+    'unicorn/numeric-separators-style': 'error',
+    'unicorn/prefer-add-event-listener': 'error',
+    'unicorn/prefer-array-find': 'error',
+    'unicorn/prefer-array-flat': 'error',
+    'unicorn/prefer-array-flat-map': 'error',
+    'unicorn/prefer-array-index-of': 'error',
+    'unicorn/prefer-array-some': 'error',
+    'unicorn/prefer-date-now': 'error',
+    'unicorn/prefer-default-parameters': 'error',
+    'unicorn/prefer-dom-node-append': 'error',
+    'unicorn/prefer-dom-node-dataset': 'error',
+    'unicorn/prefer-dom-node-remove': 'error',
+    'unicorn/prefer-dom-node-text-content': 'error',
+    'unicorn/prefer-includes': 'error',
+    'unicorn/prefer-keyboard-event-key': 'error',
+    'unicorn/prefer-math-trunc': 'error',
+    'unicorn/prefer-modern-dom-apis': 'error',
+    'unicorn/prefer-module': 'off',
+    'unicorn/prefer-negative-index': 'error',
+    'unicorn/prefer-node-protocol': 'off',
+    'unicorn/prefer-number-properties': 'error',
+    'unicorn/prefer-optional-catch-binding': 'error',
+    'unicorn/prefer-reflect-apply': 'error',
+    'unicorn/prefer-regexp-test': 'error',
+    'unicorn/prefer-set-has': 'error',
+    'unicorn/prefer-spread': 'error',
+    'unicorn/prefer-string-replace-all': 'off',
+    'unicorn/prefer-string-slice': 'error',
+    'unicorn/prefer-string-starts-ends-with': 'error',
+    'unicorn/prefer-string-trim-start-end': 'error',
+    'unicorn/prefer-switch': 'error',
+    'unicorn/prefer-ternary': 'error',
+    'unicorn/prefer-type-error': 'error',
+    'unicorn/prevent-abbreviations': 'off',
+    'unicorn/string-content': 'off',
+    'unicorn/throw-new-error': 'error',
+  },
 };
