@@ -3,8 +3,15 @@ import { createParamDecorator } from '@nestjs/common';
 
 // eslint-disable-next-line @typescript-eslint/naming-convention
 export const AuthUser = createParamDecorator(
-  (_data: unknown, ctx: ExecutionContext) => {
-    const request = ctx.switchToHttp().getRequest();
-    return request.user;
+  (_data: unknown, context: ExecutionContext) => {
+    const request = context.switchToHttp().getRequest();
+
+    const user = request.user;
+
+    if (user?.[Symbol.for('isPublic')]) {
+      return;
+    }
+
+    return user;
   },
 );
