@@ -1,3 +1,4 @@
+/* eslint-disable @typescript-eslint/no-non-null-assertion */
 import type {
   EntitySubscriberInterface,
   InsertEvent,
@@ -21,8 +22,10 @@ export class UserSubscriber implements EntitySubscriberInterface<UserEntity> {
   }
 
   beforeUpdate(event: UpdateEvent<UserEntity>): void {
-    if (event.entity.password !== event.databaseEntity.password) {
-      event.entity.password = UtilsProvider.generateHash(event.entity.password);
+    if (event.entity!.password !== event.databaseEntity.password) {
+      event.entity!.password = UtilsProvider.generateHash(
+        event.entity!.password,
+      );
     }
   }
 }
