@@ -3,6 +3,6 @@ import { bootstrap } from './main';
 void bootstrap().then((app) => {
   if (module.hot) {
     module.hot.accept();
-    module.hot.dispose(() => app.close());
+    module.hot.dispose(() => void app.close());
   }
 });
