@@ -2,7 +2,10 @@ import { Injectable, UnauthorizedException } from '@nestjs/common';
 import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 
+import type { RoleType } from '../../common/constants/role-type';
+import { TokenType } from '../../common/constants/token-type';
 import { ApiConfigService } from '../../shared/services/api-config.service';
+import type { UserEntity } from '../user/user.entity';
 import { UserService } from '../user/user.service';
 
 @Injectable()
@@ -17,16 +20,24 @@ export class JwtStrategy extends PassportStrategy(Strategy) {
     });
   }
 
-  async validate({ iat, exp, id: userId }) {
-    const timeDiff = exp - iat;
-    if (timeDiff <= 0) {
+  async validate(args: {
+    userId: string;
+    role: RoleType;
+    type: TokenType;
+  }): Promise<UserEntity> {
+    if (args.type !== TokenType.ACCESS_TOKEN) {
       throw new UnauthorizedException();
     }
-    const user = await this.userService.findOne(userId);
+
+    const user = await this.userService.findOne({
+      id: args.userId,
+      role: args.role,
+    });
 
     if (!user) {
       throw new UnauthorizedException();
     }
+
     return user;
   }
 }
