@@ -1,9 +1,11 @@
-import { Controller, Get, Inject } from '@nestjs/common';
+import { Controller, Get, Inject, Optional } from '@nestjs/common';
 import { ClientProxy } from '@nestjs/microservices';
 
 @Controller('posts')
 export class PostController {
-  constructor(@Inject('NATS_SERVICE') private client: ClientProxy) {}
+  constructor(
+    @Optional() @Inject('NATS_SERVICE') private client: ClientProxy,
+  ) {}
 
   @Get('search')
   call() {
