@@ -1,35 +1,37 @@
-import { ApiPropertyOptional } from '@nestjs/swagger';
+import { ApiProperty, ApiPropertyOptional } from '@nestjs/swagger';
 
 import { RoleType } from '../../../common/constants/role-type';
 import { AbstractDto } from '../../../common/dto/abstract.dto';
 import type { UserEntity } from '../user.entity';
 
+export type UserDtoOptions = Partial<{ isActive: boolean }>;
+
 export class UserDto extends AbstractDto {
   @ApiPropertyOptional()
-  firstName: string;
+  firstName?: string;
 
   @ApiPropertyOptional()
-  lastName: string;
+  lastName?: string;
 
-  @ApiPropertyOptional()
+  @ApiProperty()
   username: string;
 
   @ApiPropertyOptional({ enum: RoleType })
   role: RoleType;
 
   @ApiPropertyOptional()
-  email: string;
+  email?: string;
 
   @ApiPropertyOptional()
-  avatar: string;
+  avatar?: string;
 
   @ApiPropertyOptional()
-  phone: string;
+  phone?: string;
 
   @ApiPropertyOptional()
-  isActive: boolean;
+  isActive?: boolean;
 
-  constructor(user: UserEntity, options?: Partial<{ isActive: boolean }>) {
+  constructor(user: UserEntity, options?: UserDtoOptions) {
     super(user);
     this.firstName = user.firstName;
     this.lastName = user.lastName;
