@@ -2,34 +2,35 @@ import { Column, Entity } from 'typeorm';
 
 import { AbstractEntity } from '../../common/abstract.entity';
 import { RoleType } from '../../common/constants/role-type';
+import { UseDto } from '../../decorators/use-dto.decorator';
 import { VirtualColumn } from '../../decorators/virtual-column.decorator';
+import type { UserDtoOptions } from './dto/user-dto';
 import { UserDto } from './dto/user-dto';
 
 @Entity({ name: 'users' })
-export class UserEntity extends AbstractEntity<UserDto> {
+@UseDto(UserDto)
+export class UserEntity extends AbstractEntity<UserDto, UserDtoOptions> {
   @Column({ nullable: true })
-  firstName: string;
+  firstName?: string;
 
   @Column({ nullable: true })
-  lastName: string;
+  lastName?: string;
 
   @Column({ type: 'enum', enum: RoleType, default: RoleType.USER })
   role: RoleType;
 
   @Column({ unique: true, nullable: true })
-  email: string;
+  email?: string;
 
   @Column({ nullable: true })
-  password: string;
+  password?: string;
 
   @Column({ nullable: true })
-  phone: string;
+  phone?: string;
 
   @Column({ nullable: true })
-  avatar: string;
+  avatar?: string;
 
   @VirtualColumn()
-  fullName: string;
-
-  dtoClass = UserDto;
+  fullName?: string;
 }
