@@ -4,8 +4,10 @@ import type { UserEntity } from '../modules/user/user.entity';
 
 export class ContextProvider {
   private static readonly nameSpace = 'request';
-  private static authUserKey = 'user_key';
-  private static languageKey = 'language_key';
+
+  private static readonly authUserKey = 'user_key';
+
+  private static readonly languageKey = 'language_key';
 
   private static get<T>(key: string): T {
     return requestContext.get(ContextProvider.getKeyWithNamespace(key));
