@@ -1,35 +1,8 @@
 import bcrypt from 'bcrypt';
 
+import type { Optional } from '../types';
+
 export class UtilsProvider {
-  /**
-   * convert entity to dto class instance
-   * @param {{new(entity: E, options: any): T}} model
-   * @param {E[] | E} entity
-   * @param options
-   * @returns {T[] | T}
-   */
-  public static toDto<T, E>(
-    model: new (entity: E, options?: GetConstructorArgs<T>[1]) => T,
-    entity: E,
-    options?: GetConstructorArgs<T>[1],
-  ): T;
-  public static toDto<T, E>(
-    model: new (entity: E, options?: GetConstructorArgs<T>[1]) => T,
-    entity: E[],
-    options?: GetConstructorArgs<T>[1],
-  ): T[];
-  public static toDto<T, E>(
-    model: new (entity: E, options?: GetConstructorArgs<T>[1]) => T,
-    entity: E | E[],
-    options?: GetConstructorArgs<T>[1],
-  ): T | T[] {
-    if (Array.isArray(entity)) {
-      return entity.map((u) => new model(u, options));
-    }
-
-    return new model(entity, options);
-  }
-
   /**
    * generate hash from password or string
    * @param {string} password
@@ -49,16 +22,21 @@ export class UtilsProvider {
       .replace(/[^\dA-Za-z]+/g, '')
       .slice(0, Math.max(0, length));
   }
+
   /**
    * validate text with hash
    * @param {string} password
    * @param {string} hash
    * @returns {Promise<boolean>}
    */
-  static validateHash(password: string, hash: string): Promise<boolean> {
+  static validateHash(
+    password: string,
+    hash: Optional<string>,
+  ): Promise<boolean> {
     if (!password || !hash) {
       return Promise.resolve(false);
     }
+
     return bcrypt.compare(password, hash);
   }
 }
