@@ -5,12 +5,13 @@ import type { translateOptions } from 'nestjs-i18n/dist/services/i18n.service';
 
 import { AbstractDto } from '../../common/dto/abstract.dto';
 import { TRANSLATION_DECORATOR_KEY } from '../../decorators/translate.decorator';
-import type { ITranslationDecoratorInterface } from '../../interfaces/ITranslationDecoratorInterface';
+import type { ITranslationDecoratorInterface } from '../../interfaces';
 import { ContextProvider } from '../../providers/context.provider';
 
 @Injectable()
 export class TranslationService {
   constructor(private readonly i18n: I18nService) {}
+
   async translate(
     key: string,
     options: translateOptions = {},
@@ -24,18 +25,22 @@ export class TranslationService {
         if (isString(value)) {
           const translateDec: ITranslationDecoratorInterface =
             Reflect.getMetadata(TRANSLATION_DECORATOR_KEY, dto, key);
+
           if (translateDec.translationKey) {
             await this.translate(`${translateDec.translationKey}.${value}`, {
               lang: ContextProvider.getLanguage(),
             });
           }
+
           return;
         }
 
         if (value instanceof AbstractDto) {
           await this.translateNecessaryKeys(value);
+
           return;
         }
+
         if (isArray(value)) {
           await Promise.all(
             map(value, (v) => {
@@ -47,6 +52,7 @@ export class TranslationService {
         }
       }),
     );
+
     return dto;
   }
 }
