@@ -1,6 +1,7 @@
-import { Global, HttpModule, Module } from '@nestjs/common';
-import { ClientProxyFactory, Transport } from '@nestjs/microservices';
+import { HttpModule } from '@nestjs/axios';
+import { Global, Module } from '@nestjs/common';
 
+// import { ClientProxyFactory, Transport } from '@nestjs/microservices';
 import { ApiConfigService } from './services/api-config.service';
 import { AwsS3Service } from './services/aws-s3.service';
 import { GeneratorService } from './services/generator.service';
@@ -13,20 +14,22 @@ const providers = [
   AwsS3Service,
   GeneratorService,
   TranslationService,
-  {
-    provide: 'NATS_SERVICE',
-    useFactory: (configService: ApiConfigService) => {
-      const natsConfig = configService.natsConfig;
-      return ClientProxyFactory.create({
-        transport: Transport.NATS,
-        options: {
-          name: 'NATS_SERVICE',
-          url: `nats://${natsConfig.host}:${natsConfig.port}`,
-        },
-      });
-    },
-    inject: [ApiConfigService],
-  },
+
+
+  // {
+  //   provide: 'NATS_SERVICE',
+  //   useFactory: (configService: ApiConfigService) => {
+  //     const natsConfig = configService.natsConfig;
+  //     return ClientProxyFactory.create({
+  //       transport: Transport.NATS,
+  //       options: {
+  //         name: 'NATS_SERVICE',
+  //         url: `nats://${natsConfig.host}:${natsConfig.port}`,
+  //       },
+  //     });
+  //   },
+  //   inject: [ApiConfigService],
+  // },
 ];
 
 @Global()
